# perf: use class instead of object literals with getters