--- a/benchmarks/fetch/request-creation.mjs
+++ b/benchmarks/fetch/request-creation.mjs
@@ -1,0 +1,8 @@
+import { bench, run } from 'mitata'
+import { Request } from '../../lib/web/fetch/request.js'
+
+const input = 'https://example.com/post'
+
+bench('new Request(input)', () => new Request(input, undefined))
+
+await run()