--- a/lib/web/fetch/request.js
+++ b/lib/web/fetch/request.js
@@ -41,6 +41,21 @@ const requestFinalizer = new FinalizationRegistry(({ signal, abort }) => {
 
 let patchMethodWarning = false
 
+class RequestRealmSettingsObject {
+  baseUrl = getGlobalOrigin()
+
+  get origin () {
+    return this.baseUrl?.origin
+  }
+
+  policyContainer = makePolicyContainer()
+
+  static {
+    delete this.prototype.constructor
+    Object.freeze(this.prototype)
+  }
+}
+
 // https://fetch.spec.whatwg.org/#request-class
 class Request {
   // https://fetch.spec.whatwg.org/#dom-request
@@ -56,13 +71,9 @@ class Request {
 
     // https://html.spec.whatwg.org/multipage/webappapis.html#environment-settings-object
     this[kRealm] = {
-      settingsObject: {
-        baseUrl: getGlobalOrigin(),
-        get origin () {
-          return this.baseUrl?.origin
-        },
-        policyContainer: makePolicyContainer()
-      }
+      // Note: Slow initialization of object literals with getters.
+      // TODO: Remove workaround.
+      settingsObject: new RequestRealmSettingsObject()
     }
 
     // 1. Let request be null.