--- a/lib/web/fetch/request.js
+++ b/lib/web/fetch/request.js
@@ -49,11 +49,6 @@ class RequestRealmSettingsObject {
   }
 
   policyContainer = makePolicyContainer()
-
-  static {
-    delete this.prototype.constructor
-    Object.freeze(this.prototype)
-  }
 }
 
 // https://fetch.spec.whatwg.org/#request-class