--- a/lib/web/fetch/request.js
+++ b/lib/web/fetch/request.js
@@ -41,7 +41,7 @@ const requestFinalizer = new FinalizationRegistry(({ signal, abort }) => {
 
 let patchMethodWarning = false
 
-class RequestRealmSettingsObject {
+class EnvironmentSettingsObjectBase {
   baseUrl = getGlobalOrigin()
 
   get origin () {
@@ -49,6 +49,10 @@ class RequestRealmSettingsObject {
   }
 
   policyContainer = makePolicyContainer()
+}
+
+class EnvironmentSettingsObject {
+  settingsObject = new EnvironmentSettingsObjectBase()
 }
 
 // https://fetch.spec.whatwg.org/#request-class
@@ -65,12 +69,9 @@ class Request {
     init = webidl.converters.RequestInit(init)
 
     // https://html.spec.whatwg.org/multipage/webappapis.html#environment-settings-object
-    this[kRealm] = {
-      // Note: Slow initialization of object literals with getters.
-      // TODO: Remove workaround.
-      settingsObject: new RequestRealmSettingsObject()
-    }
-
+    // Note: Slow initialization of object literals with getters.
+    // TODO: Remove workaround.
+    this[kRealm] = new EnvironmentSettingsObject()
     // 1. Let request be null.
     let request = null