--- a/lib/web/fetch/request.js
+++ b/lib/web/fetch/request.js
@@ -373,7 +373,6 @@ class Request {
     // (https://dom.spec.whatwg.org/#dom-abortsignal-any)
     const ac = new AbortController()
     this[kSignal] = ac.signal
-    this[kSignal][kRealm] = this[kRealm]
 
     // 29. If signal is not null, then make this’s signal follow signal.
     if (signal != null) {
@@ -443,7 +442,6 @@ class Request {
     this[kHeaders] = new Headers(kConstruct)
     this[kHeaders][kHeadersList] = request.headersList
     this[kHeaders][kGuard] = 'request'
-    this[kHeaders][kRealm] = this[kRealm]
 
     // 31. If this’s request’s mode is "no-cors", then:
     if (mode === 'no-cors') {
@@ -888,11 +886,9 @@ function fromInnerRequest (innerRequest, signal, guard, realm) {
   request[kState] = innerRequest
   request[kRealm] = realm
   request[kSignal] = signal
-  request[kSignal][kRealm] = realm
   request[kHeaders] = new Headers(kConstruct)
   request[kHeaders][kHeadersList] = innerRequest.headersList
   request[kHeaders][kGuard] = guard
-  request[kHeaders][kRealm] = realm
   return request
 }