--- a/lib/web/fetch/request.js
+++ b/lib/web/fetch/request.js
@@ -70,7 +70,6 @@ class Request {
 
     // https://html.spec.whatwg.org/multipage/webappapis.html#environment-settings-object
     // Note: Slow initialization of object literals with getters.
-    // TODO: Remove workaround.
     this[kRealm] = new EnvironmentSettingsObject()
     // 1. Let request be null.
     let request = null