--- a/lib/web/fetch/response.js
+++ b/lib/web/fetch/response.js
@@ -138,7 +138,6 @@ class Response {
     this[kHeaders] = new Headers(kConstruct)
     this[kHeaders][kGuard] = 'response'
     this[kHeaders][kHeadersList] = this[kState].headersList
-    this[kHeaders][kRealm] = this[kRealm]
 
     // 3. Let bodyWithType be null.
     let bodyWithType = null
@@ -522,7 +521,6 @@ function fromInnerResponse (innerResponse, guard, realm) {
   response[kHeaders] = new Headers(kConstruct)
   response[kHeaders][kHeadersList] = innerResponse.headersList
   response[kHeaders][kGuard] = guard
-  response[kHeaders][kRealm] = realm
   return response
 }