--- a/lib/web/fetch/response.js
+++ b/lib/web/fetch/response.js
@@ -29,12 +29,17 @@ const { types } = require('node:util')
 
 const textEncoder = new TextEncoder('utf-8')
 
+class EnvironmentSettingsObject {
+  settingsObject = {}
+}
+
+
 // https://fetch.spec.whatwg.org/#response-class
 class Response {
   // Creates network error Response.
   static error () {
     // TODO
-    const relevantRealm = { settingsObject: {} }
+    const relevantRealm = new EnvironmentSettingsObject()
 
     // The static error() method steps are to return the result of creating a
     // Response object, given a new network error, "immutable", and this’s
@@ -62,7 +67,7 @@ class Response {
 
     // 3. Let responseObject be the result of creating a Response object, given a new response,
     //    "response", and this’s relevant Realm.
-    const relevantRealm = { settingsObject: {} }
+    const relevantRealm = new EnvironmentSettingsObject()
     const responseObject = fromInnerResponse(makeResponse({}), 'response', relevantRealm)
 
     // 4. Perform initialize a response given responseObject, init, and (body, "application/json").
@@ -74,7 +79,7 @@ class Response {
 
   // Creates a redirect Response that redirects to url with status status.
   static redirect (url, status = 302) {
-    const relevantRealm = { settingsObject: {} }
+    const relevantRealm = new EnvironmentSettingsObject()
 
     webidl.argumentLengthCheck(arguments, 1, { header: 'Response.redirect' })
 
@@ -127,7 +132,7 @@ class Response {
     init = webidl.converters.ResponseInit(init)
 
     // TODO
-    this[kRealm] = { settingsObject: {} }
+    this[kRealm] = new EnvironmentSettingsObject()
 
     // 1. Set this’s response to a new response.
     this[kState] = makeResponse({})