--- a/lib/web/fetch/response.js
+++ b/lib/web/fetch/response.js
@@ -33,7 +33,6 @@ class EnvironmentSettingsObject {
   settingsObject = {}
 }
 
-
 // https://fetch.spec.whatwg.org/#response-class
 class Response {
   // Creates network error Response.