--- a/lib/web/fetch/response.js
+++ b/lib/web/fetch/response.js
@@ -12,7 +12,8 @@ const {
   isBlobLike,
   serializeJavascriptValueToJSONString,
   isErrorLike,
-  isomorphicEncode
+  isomorphicEncode,
+  EnvironmentSettingsObject
 } = require('./util')
 const {
   redirectStatusSet,
@@ -29,15 +30,10 @@ const { types } = require('node:util')
 
 const textEncoder = new TextEncoder('utf-8')
 
-class EnvironmentSettingsObject {
-  settingsObject = {}
-}
-
 // https://fetch.spec.whatwg.org/#response-class
 class Response {
   // Creates network error Response.
   static error () {
-    // TODO
     const relevantRealm = new EnvironmentSettingsObject()
 
     // The static error() method steps are to return the result of creating a