--- a/lib/web/fetch/response.js
+++ b/lib/web/fetch/response.js
@@ -22,7 +22,6 @@ const {
 const { kState, kHeaders, kGuard, kRealm } = require('./symbols')
 const { webidl } = require('./webidl')
 const { FormData } = require('./formdata')
-const { getGlobalOrigin } = require('./global')
 const { URLSerializer } = require('./data-url')
 const { kHeadersList, kConstruct } = require('../../core/symbols')
 const assert = require('node:assert')
@@ -87,7 +86,7 @@ class Response {
     // TODO: base-URL?
     let parsedURL
     try {
-      parsedURL = new URL(url, getGlobalOrigin())
+      parsedURL = new URL(url, relevantRealm.settingsObject.baseUrl)
     } catch (err) {
       throw new TypeError(`Failed to parse URL from ${url}`, { cause: err })
     }