--- a/lib/web/fetch/util.js
+++ b/lib/web/fetch/util.js
@@ -1572,11 +1572,11 @@ function utf8DecodeBytes (buffer) {
 class EnvironmentSettingsObjectBase {
   #baseUrl
 
-  get baseUrl() {
+  get baseUrl () {
     return (this.#baseUrl ??= getGlobalOrigin())
   }
 
-  set baseUrl(baseUrl) {
+  set baseUrl (baseUrl) {
     this.#baseUrl = baseUrl
   }