--- a/lib/web/fetch/util.js
+++ b/lib/web/fetch/util.js
@@ -1570,15 +1570,7 @@ function utf8DecodeBytes (buffer) {
 }
 
 class EnvironmentSettingsObjectBase {
-  #baseUrl
-
-  get baseUrl () {
-    return (this.#baseUrl ??= getGlobalOrigin())
-  }
-
-  set baseUrl (baseUrl) {
-    this.#baseUrl = baseUrl
-  }
+  baseUrl =  getGlobalOrigin()
 
   get origin () {
     return this.baseUrl?.origin