--- a/lib/web/fetch/util.js
+++ b/lib/web/fetch/util.js
@@ -1570,7 +1570,7 @@ function utf8DecodeBytes (buffer) {
 }
 
 class EnvironmentSettingsObjectBase {
-  baseUrl =  getGlobalOrigin()
+  baseUrl = getGlobalOrigin()
 
   get origin () {
     return this.baseUrl?.origin