Slow initialization of object literals with getters.

- main
```
benchmark               time (avg)             (min … max)       p75       p99      p999
---------------------------------------------------------- -----------------------------
new Request(input)  12'398 ns/iter   (8'100 ns … 4'134 µs) 11'000 ns 34'100 ns    147 µs
```

- this patch

```
benchmark               time (avg)             (min … max)       p75       p99      p999
---------------------------------------------------------- -----------------------------
new Request(input)  10'997 ns/iter   (7'000 ns … 3'704 µs)  9'700 ns 34'000 ns    159 µs
```