@Uzlopak 2024-04-19
Did you also see that makePolicyContainer() seemed to be super slow?