@KhafraDev 2024-04-19
> Did you also see that makePolicyContainer() seemed to be super slow?

It returns an object with a single property. If you're running creating a hundred thousand Requests everything will seem to be "super slow".