@mcollina 2024-04-20
There is nothing absolutely “slow” or “fast”, as those are relative concepts. The only useful measure when doing optimization work is to identify bottlenecks and remove them.

(A tiny fast function called a bazillion times might be causing a bottleneck as much as a huge slow function seldomly called).