@KhafraDev 2024-04-20
not in this context, it's called once per Request and is needed for spec compatibility