@tsctx 2024-04-20
> not in this context, it's called once per Request and is needed for spec compatibility

What do you mean by that?  I've already created a new context for each call.