@KhafraDev 2024-04-20
EnvironmentSettingsObject should be shared between Response and Request, it's meant to be the same object. It's fine to lazily create the properties needed for each.