@tsctx 2024-04-21
> I guess that can lead to further perf improvements, the `realm` should be the exact same object and it can't change because they're non-transferable and cannot be cloned. The issue there being that the global origin may be mutated, but that can be a getter while policyContainer and whatever else is constant?

If we use getter, there should be no problem.