@KhafraDev 2024-04-21
I don't think that those changes will solve the issue, consider the following:

```mjs
import { setGlobalOrigin, Response } from 'undici'

setGlobalOrigin('http://localhost:3000')
new Request('/').url // http://localhost:3000

setGlobalOrigin('https://www.google.com')
new Request('/').url // https://www.google.com
```

If this works, please add a test so we don't regress in the future.